Allow user to locate wrapped text lines
This little changes adds an 'eachLine' option to 'text'. You can set
it to a callback that will be called with the arguments (text, x, y)
for each line of text that gets generated.

This makes it possible to figure out eactly where the text has been
placed in order to add annotations to it, or add other graphics that
need to be placed relative to the text.