bugfix with text align in rtl languages
I have added an attribute for `options` called `dir`, that defaults to `ltr`.
if user set that option to `rtl` when using `text` method we needed to change
the
default text align to `right`. In `justify` mode we needed to align the last
paragraph to the `right`.

Still we have a bug with calculating the correct line width for Arabic texts.
seems the calculation does not take account the Arabic characters being
reshaped. (the calculation is not correct but character reshaping is done
right).

Also when the text is aligned to right and its direction is `rtl`, to calculate
the line width we need to remove the spaces from the beginning of text not the
end of it. so I changed code accordingly. without this changed right aligned
texts in `rtl` mode always had a padding from right side.